Backuptool: remove hosts backup/restore script
The 'hardened microG' build flavor delivers a monthly updated hosts file
with many ad-blocking settings. A backup/restore of the hosts file works
against this approach.

As the only remaining "root app" solution to e.g. allow a hosts editor
is Magisk, which implements this feature 'system-less'ly, the extended
backup/restore option, which was previously used in the 15.1...18.1
microG builds has not been ported, as it does not really make sense any
longer.

Change-Id: Idecf8091baab9c7619123ab2f2373154c52fe7e8